+++ src/main/evaluation/pattern.scala
@@ -4,11 +4,11 @@ import translation.structure.*
 import evaluation.expression.evaluateLiteralExpr
 import evaluation.config.*
 
-def evaluatePattern(pattern: Pattern): Int = pattern match {
+def evaluatePattern(pattern: Pattern): Double = pattern match {
   case UnNamed          => UnNamedPatternCost
   case Constant(value)  => ConstantPatternCost + evaluateLiteralExpr(value)
   case Basic(_, values) =>
-    BasicPatternCost + values.fold(0)(_.map(evaluatePattern).sum)
+    BasicPatternCost + values.fold(0.0)(_.map(evaluatePattern).sum)
   case PatternTuple(elements) =>
     PatternTupleCost + elements.map(evaluatePattern).sum
 }
